admin/user 네이밍수정-김다율

+++ src/main/java/com/green/babyfood/product/ProductController.java
@@ -22,7 +22,7 @@ public class ProductController {
     @GetMapping
     @Operation(summary = "상품 상세보기 페이지",description = ""+
             "productId = 상품 코드")
-    ProductSelDto selProduct(int productId){
+    ProductSelDto selProduct(Long productId){
         return service.selProduct(productId);
     }
